requirements-dev.txt
pep8
avakas
nose2
shyaml
awscli
pylint
future
bandit
requests_mock
vulture
flexmock
